requirements.txt
tinydb==3.7.0
speedtest_cli==1.0.7
